Move chart mode files into airspace-3d, layer-toggle, and interaction subdirs
