+++ apps/atlas/src/modes/chart/interaction/click-to-select.spec.ts
@@ -1,21 +1,22 @@
 import { describe, it, expect } from 'vitest';
 
+import { AIRPORTS_LAYER_ID } from '../layers/airports-layer.tsx';
+import {
+  AIRSPACE_FILL_EXTRUSION_LAYER_ID,
+  AIRSPACE_FILL_LAYER_ID,
+  AIRSPACE_LINE_LAYER_ID,
+} from '../layers/airspace-layer.tsx';
+import { AIRWAYS_LAYER_ID } from '../layers/airways-layer.tsx';
+import { FIXES_LAYER_ID } from '../layers/fixes-layer.tsx';
+import { NAVAIDS_LAYER_ID } from '../layers/navaids-layer.tsx';
+
+import type { InspectableFeature } from './click-to-select.ts';
 import {
   classifyClick,
   formatChipLabel,
   pickFeatureByPriority,
   selectedFromFeature,
 } from './click-to-select.ts';
-import type { InspectableFeature } from './click-to-select.ts';
-import { AIRPORTS_LAYER_ID } from './layers/airports-layer.tsx';
-import {
-  AIRSPACE_FILL_EXTRUSION_LAYER_ID,
-  AIRSPACE_FILL_LAYER_ID,
-  AIRSPACE_LINE_LAYER_ID,
-} from './layers/airspace-layer.tsx';
-import { AIRWAYS_LAYER_ID } from './layers/airways-layer.tsx';
-import { FIXES_LAYER_ID } from './layers/fixes-layer.tsx';
-import { NAVAIDS_LAYER_ID } from './layers/navaids-layer.tsx';
 
 /**
  * Builds the minimal `InspectableFeature` shape `selectedFromFeature`
