+++ apps/atlas/src/modes/chart/interaction/disambiguation-popover.spec.tsx
@@ -2,14 +2,15 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import type { ReactElement, ReactNode } from 'react';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
+import { HighlightProvider } from '../highlight-provider.tsx';
+import { AIRPORTS_LAYER_ID } from '../layers/airports-layer.tsx';
+import { AIRSPACE_FILL_LAYER_ID, AIRSPACE_LINE_LAYER_ID } from '../layers/airspace-layer.tsx';
+import { AIRWAYS_LAYER_ID } from '../layers/airways-layer.tsx';
+import { FIXES_LAYER_ID } from '../layers/fixes-layer.tsx';
+import { NAVAIDS_LAYER_ID } from '../layers/navaids-layer.tsx';
+
 import type { InspectableFeature } from './click-to-select.ts';
 import { DisambiguationPopover } from './disambiguation-popover.tsx';
-import { HighlightProvider } from './highlight-provider.tsx';
-import { AIRPORTS_LAYER_ID } from './layers/airports-layer.tsx';
-import { AIRSPACE_FILL_LAYER_ID, AIRSPACE_LINE_LAYER_ID } from './layers/airspace-layer.tsx';
-import { AIRWAYS_LAYER_ID } from './layers/airways-layer.tsx';
-import { FIXES_LAYER_ID } from './layers/fixes-layer.tsx';
-import { NAVAIDS_LAYER_ID } from './layers/navaids-layer.tsx';
 
 const { useMapMock, onMock, offMock } = vi.hoisted(() => ({
   useMapMock: vi.fn(),
