+++ apps/atlas/src/modes/chart/layer-toggle/layer-toggle-icons.tsx
